src/test/java/com/biblioteca/repository/BookRepositoryTest.java
package com.biblioteca.repository;

import com.biblioteca.Application;
import com.biblioteca.model.Book;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.context.junit4.SpringRunner;
import org.springframework.transaction.annotation.Transactional;

import java.util.List;

import static org.junit.Assert.assertEquals;

@Transactional
@RunWith(SpringRunner.class)
@SpringBootTest(classes = Application.class)
public class BookRepositoryTest {

    @Autowired
    private BookRepository bookRepository;

    @Test
    public void shouldReturnListOfBooks() {
        List<Book> books = bookRepository.getAllBooks();
        assertEquals(books.size(), 4);
    }

}
